+++ backend/api/models/__init__.py
@@ -1,0 +1,5 @@
+from .comment import Comment
+from .exam import ProficiencyExam, Exam
+from .language import Language
+from .question import Question, QuestionOption
+from .user import User
