+++ backend/api/serializers/__init__.py
@@ -1,0 +1,5 @@
+from .language import LanguageSerializer
+from .question import QuestionOptionSerializer, QuestionSerializer
+from .exam import ProficiencyExamSerializer
+from .user import ProfileSerializer
+from .comment import CommentSerializer
