+++ backend/api/serializers/question.py
@@ -1,0 +1,17 @@
+from rest_framework import serializers
+
+from ..models import QuestionOption, Question
+
+
+class QuestionOptionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = QuestionOption
+        fields = ('text',)
+
+
+class QuestionSerializer(serializers.ModelSerializer):
+    question_options = QuestionOptionSerializer(many=True)
+
+    class Meta:
+        model = Question
+        fields = ('id', 'text', 'question_options')
