+++ backend/api/test/__init__.py
@@ -1,0 +1,3 @@
+from .exam import ProficiencyTest
+from .user import LoginViewTests, RegisterViewTests, ProfileViewTests
+
