+++ backend/api/test/user.py
@@ -1,67 +1,7 @@
-from rest_framework.test import APITestCase, APIClient
 from rest_framework import status
+from rest_framework.test import APITestCase, APIClient
 
-
-from api.models.comment import Comment
-from api.models.language import Language
-from api.models.user import User
-
-"""
-class ProficiencyTest(APITestCase):
-    def setUp(self):
-        q = {
-            "questionLevel": "A2",
-            "questionType": "vocab",
-            "questionId": "asd12ejd",
-            "questionAnswer": "bring",
-            "questionText": "to take or carry someone or \
-                something to a place or a person, or \
-                in the direction of the person speaking",
-            "questionOptions": [
-                "bring",
-                "catch",
-                "follow",
-                "burn"
-            ],
-        }
-        question = Question.objects.create(**q)
-
-        self.req = {
-            "testLanguage": "english",
-        }
-
-    def test_number_of_question(self):
-        
-        response = self.client.post('/proficiency', self.req)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data.get('testQuestions')), 6)
-
-    def test_number_of_question_options(self):
-        
-        response = self.client.post('/proficiency', self.req)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        numberOfQuestions = len(response.data.get('testQuestions'))
-        for i in range(numberOfQuestions):
-            self.assertEqual(
-                len(response.data.get('testQuestions')[i].get('questionOptions')),
-                4)
-
-    def test_answer_in_options(self):
-
-        response = self.client.post('/proficiency', self.req)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        numberOfQuestions = len(response.data.get('testQuestions'))
-        for i in range(numberOfQuestions):
-            options = response.data.get('testQuestions')[i].get('questionOptions')
-            answer = response.data.get('testQuestions')[i].get('questionAnswer')
-            self.assertTrue(answer in options)
-            options.remove(answer)
-            self.assertTrue(not answer in options)
-  
-
-"""
+from ..models import User, Language, Comment
 
 
 class LoginViewTests(APITestCase):
@@ -76,7 +16,7 @@ class LoginViewTests(APITestCase):
         user.set_password('isa21-ad')
         user.save()
 
-    def test_succesful_email(self):
+    def test_successful_email(self):
         req = {
             'email_username': 'adaLovelace@email.com',
             'password': 'isa21-ad'
@@ -102,7 +42,7 @@ class LoginViewTests(APITestCase):
         response = self.client.post('/user/login', req)
 
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-        self.assertTrue(not 'token' in response.data.keys())
+        self.assertTrue('token' not in response.data.keys())
 
     def test_missing_field(self):
         req = {
@@ -114,11 +54,11 @@ class LoginViewTests(APITestCase):
             del deficient_req[key]
             response = self.client.post('/user/login', deficient_req)
             self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-            self.assertTrue(not 'token' in response.data.keys())
+            self.assertTrue('token' not in response.data.keys())
 
         response = self.client.post('/user/login', {})
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertTrue(not 'token' in response.data.keys())
+        self.assertTrue('token' not in response.data.keys())
 
     def test_wrong_password(self):
         req = {
@@ -127,11 +67,11 @@ class LoginViewTests(APITestCase):
         }
         response = self.client.post('/user/login', req)
         self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertTrue(not 'token' in response.data.keys())
+        self.assertTrue('token' not in response.data.keys())
 
 
 class RegisterViewTests(APITestCase):
-    def test_succesful_register(self):
+    def test_successful_register(self):
         req = {
             "name": "Ada",
             "surname": "Lovelace",
@@ -158,18 +98,13 @@ class RegisterViewTests(APITestCase):
             del deficient_req[key]
             response = self.client.post('/user/register', deficient_req)
             self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-            self.assertTrue(not 'token' in response.data.keys())
+            self.assertTrue('token' not in response.data.keys())
             self.assertTrue('message' in response.data.keys())
 
         response = self.client.post('/user/register', {})
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertTrue(not 'token' in response.data.keys())
+        self.assertTrue('token' not in response.data.keys())
         self.assertTrue('message' in response.data.keys())
-
-
-class GuestViewTests(APITestCase):
-    def test_is_token_valid(self):
-        pass
 
 
 class ProfileViewTests(APITestCase):
@@ -208,9 +143,9 @@ class ProfileViewTests(APITestCase):
         user.save()
 
         data = {
-            'username':'ada21',
-            'comment':'good user',
-            'rate':4,
+            'username': 'ada21',
+            'comment': 'good user',
+            'rate': 4,
         }
 
         comment = Comment(**data, commented_user=user)
@@ -229,4 +164,3 @@ class ProfileViewTests(APITestCase):
         self.client.force_authenticate(user=user)
         response = self.client.get('/user/profile?username=alan1')
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-
