+++ backend/api/views/__init__.py
@@ -1,0 +1,3 @@
+from .exam import ProficiencyView
+from .user import RegisterView, LoginView, GuestView, ProfileView
+
